trustfence-stm-signtools: include external libraries in the package itself
The signing tools have a dependency of libQt5Core.so.5, which implies that this
library must to be installed on the native PC. This commit includes all the
required shared libraries for the signing tools inside the own package to avoid
external dependencies. With this change there is not needed any more the qtbase
dependency at build time.
Package version has been bumped to 1.2.

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>