sdk: move templates to conf/templates in preparation for Yocto 5.0 (Scarthgap)
Starting in poky commit 8791c77a4098d62080ecad09d94718eccd7f1a47, there is
now a check to make sure TEMPLATECONF follows specific rules. If we don't abide
by them, an error will occur when creating a project with our mkproject.sh
script. Move the templates to the place they're supposed to be in and reflect
the change in mkproject.sh.

https://onedigi.atlassian.net/browse/DEL-9011

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>