bluez5: adapt bbappend to 5.72
This is the version of the recipe provided by poky in Yocto 5.0 scarthgap.
Aside from updating the verison number, explicitly create /etc/bluetooth
directory during installation.

The creation of this directory was removed from the recipe's base do_install()
in poky (see poky commit 55692591227eaac2d50ab339eea87ddca395f6df), so we need
to create it in our bbappend to be able to add files to it.

https://onedigi.atlassian.net/browse/DEL-9011

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>