freescale-layer: weston-init: adapt to scarthgap
A lot of the imx93-specific changes have been upstreamed to meta-freescale, and
this version is now used in all platforms, so we don't need to include
version-specific code (for example, to manage weston-socket.sh).

Also, update our do_install:append() to match the one in meta-imx, while
keeping our customizations.

https://onedigi.atlassian.net/browse/DEL-9011
https://onedigi.atlassian.net/browse/DEL-9081

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>