global: remove IMAGE_NAME_SUFFIX from IMAGE_LINK_NAME
poky has reworked the format of image names, hardcoding the IMAGE_NAME_SUFFIX
(.rootfs) into IMAGE_LINK_NAME, the name used for image symlinks. This would
only be considered a cosmetic change if it weren't for the fact that we have
scripts and test infrastructure that relies on the old image name format, and
this change would force us to rework it all.

Remove the suffix from the link name to maintain our image name format, and
reflect this name in our custom image type scripts. Note that this removes
".rootfs" from several files, for example:

    * .rootfs.ext4.gz is now .ext4.gz
    * .rootfs.cpio.gz.u-boot.tf is now .cpio.gz.u-boot.tf
    * .rootfs.ubifs is now .ubifs
    * .rootfs.sdcard is now .sdcard

However, symlink names and non-rootfs files (.boot.vfat, .recovery.vfat...)
are unchanged.

https://onedigi.atlassian.net/browse/DEL-9011

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>