wpa-supplicant: use our custom defconfig
A recent change in poky changed the configuration behavior of this package,
using the defconfig in the source tree instead of the one provided in Yocto
(which we overwrite in meta-digi). Since our ccmp1 builds have a lot of patches
specific to the platforms' wifi chip, using the default defconfig results in
linking errors.

Restore the ability to use custom defconfigs to be able to use the ones defined
in our layer per platform.

https://onedigi.atlassian.net/browse/DEL-9011

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>