imx-boot: fix compilation for imx8m platforms
In meta-freescale commit 4d64dde1686a017ebe2763dd7880563a6fc51b53,
compile_mx8m() was modified to account for possible configuration suffixes in
the dtb filename via the creation of a symlink. In our case, the filename is
the same as the target, causing the dtb to get replaced with a dead symlink.
For now, revert this function to how it was in the kirkstone branch of
meta-freescale to avoid this.

https://onedigi.atlassian.net/browse/DEL-9011
https://onedigi.atlassian.net/browse/DEL-9081

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>