stm32mpu-ai: tensorflow-lite: backport recipe from github st main
Current recipe build was failing due to dependencies with cpuinfo header files and
libraries. This commit backports the recipe that is currently being built in the
ST main branch in GitHub which clears these dependencies.

The recipe was backported as it was in commit with SHA256:
fc56201c67adf899080cff5b45ed33824e427609

https://onedigi.atlassian.net/browse/CCS-12

Signed-off-by: David Escalona <david.escalona@digi.com>