pulseaudio: ccmp25: avoid speaker mute
based on commit 951f20d36b8575e121a27e5acd7ad68c722a98af, this fix
avoids muting the speaker control.
Despite the original commit is for Headphone, it also applies, for the
ccmp25, to the Speaker control.
While on it, reword the files so the change applies to all stm platforms
by using the common machine.

https://onedigi.atlassian.net/browse/DEL-9227

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>