freescale-layer: optee-os: fix staticdev package QA errors
For some reason, using the '+=' operator instead of ':append' when specifying
the staticdev package's FILES, the original values are overwritten, causing all
of them to get included in the dev package instead. Since some of these files
are static libraries, this causes QA errors.

Replace the '+=' operator with ':append' to fix this.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>