swupdate: clean warning on firmware update
The swupdate description file allows to configure the setting of u-boot
environment variables. For some platforms the "upgrade_available" variable
has an empty value, and swupdate parser warns when the description asks
to set a variable with an empty value.
    
Initialize the variable to 0, which is semantically the same for u-boot.

https://onedigi.atlassian.net/browse/DEL-9279

Signed-off-by: Javier Viguera <javier.viguera@digi.com>