eiq-examples: fix dependence on vela native tool
The recipe level DEPENDS on ethos-u-vela-native is insufficient to ensure
the native vela tool population into the eiq-examples recipe sysroot
native directory before use. The new download_transform_models task uses
that tool, so add the dependence on the task to ensure it is available
before use.

While on it, configure the working directory of the new task using the
Yocto way (using task[dirs]).

Signed-off-by: Javier Viguera <javier.viguera@digi.com>