packagegroup-qt6-addons: only add qtpdf when "webengine" is a DISTRO_FEATURE
This was always the case until commit 498e94c9132e90fc1c406b8422e4554c6801918e
in meta-qt6, which was meant to reorganize some packages depending on the
target architecture, but removed the check for this package in the process
(while maintaining the check for other related packages). We haven't tried
building qtpdf until now, and since it's causing build errors out of the box,
we might as well remove it for now because it's only meant to be an addon
exclusive to the SDK.

Since the new package qtdoc has a dependency with qtpdf, remove it as well.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>