Move common SDK modifications to a new bbclass, dey-image-sdk
Our new dey-toolchain recipe was missing the logic needed to create the
"dey-version" file in the SDK's installation path. Move this logic from the
dey-image bbclass to a new bbclass meant exclusively for SDK logic, and inherit
it where needed. While at it, move other common SDK modifications to this
class.

https://onedigi.atlassian.net/browse/DEL-9297

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>