bluetooth-init: cc9: do not unload the btnxpuart on reboot
Do not take any action such as removing the btnxpuart module When
the system is being restarted (poweroff or reboot), as the driver
tries to restore the UART baudrate when removing the module, and if
the system is being off, some of the entries might be dissapeared in the
middle of the process, leading to a NULL pointer and preventing the
system of being rebooted.
The btnxpuart can be removed on suspend/resume safely, so add the logic
to detect if the system is being rebooted.

https://onedigi.atlassian.net/browse/DEL-9290

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>