connectcore-demo-example: restructure multimedia dependencies
At the moment, we provide four versions of the connectcore demo package:

    * webkit + multimedia: we include this one in webkit images
    * webkit (no multimedia): used in cases where the platform supports webkit,
      but video/webgl performance isn't optimal
    * regular: only contains the demo server and basic assets, we include it in
      non-webkit images
    * regular + multimedia: contains the regular package plus video/webgl web
      content

This last version doesn't make a lot of sense because video/webgl content is
meant to showcase the target's VPU/GPU capabilities, and that can only be done
when running the demo on the target via webkit. Besides, even though we don't
include this version of the package in our default non-webkit images, the
runtime dependencies in the .inc file make bitbake generate the video-examples
and webglsamples packages during the image builds, even if they don't end up
in the final image. This is true even for platforms that don't support webkit,
such as the ccimx6ul, ccimx91 and ccimx93.

Initially, I tried making the -multimedia package exclusive to the webkit
recipe, but this causes its contents to be bundled with the base (regular)
package. To avoid this, define the -multimedia package in the .inc file, and
move its dependency with video-examples and webglsamples to the webkit recipe.
This way, the regular + multimedia version of the package still exists, but
it no longer causes bitbake to generate video/webgl packages when not needed.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>