nativesdk-packagegroup-sdk-host: only include optee dependencies when needed
Both python3-cryptography and python3-pyelftools are needed to compile optee-os
with a Yocto-generated SDK, but we were adding these packages to all i.MX SDKs,
even for platforms that don't use optee. This wouldn't be a problem if it
weren't for the fact that python3-cryptography is implemented in rust, and its
inclusion in the SDKs forces bitbake to generate the rust compiler just for
that package.

Only include these python packages for platforms that use optee.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>