meta-digi: get rid of LOG git mirrors
Use the canonical Stash repository to avoid synchronization problems
between the repo and the mirror.

As that's an old VM with limited specs, there is no much gain on using
it over the canonical Stash repositories that justifies the code
overhead and the possible errors due to synchronization problems.

Manually cherry-picked from commits:
- fecc12fa0b99a5172a346dd0fdc4ea71fc2d2b8c
- b00450d8dd2d009ed1d99acaf38848aa205829ae

Signed-off-by: Javier Viguera <javier.viguera@digi.com>
Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>