weston: revert Solid Color Support
Currently, this feature prevents the CPU from properly entering the
standby state on the i.MX93.
With the current implementation, the PXP module generates
"pxp-dmaengine-std" interrupts, which prevents the CPU from asserting
the PMIC_STBY_REQ signal.
An ongoing ticket with NXP is addressing this issue.
In the meantime, use the following workaround to resolve the situation.

https://onedigi.atlassian.net/browse/DEL-9356

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>