qt6: add qt6 support for the ccmp25-dvk
Using the qt5 STM customizations as reference, adapt them to qt6 by changing
filenames and paths when needed and porting the qtbase configuration options
from the v2.1.0 branch of meta-st-x-linux-qt. Make sure that the
cinematicexperience is included in packagegroup-qt6-dey.

For now, leave the STM qt5 support as-is for backwards compatibility.

https://onedigi.atlassian.net/browse/DEL-9462

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>