stm-st-stm32mp: weston: restore wl_shell functionality to weston 13.0.1
We apply a very similar patch to the i.MX fork of weston 12.0.4, but we never
needed this patch for stm32mp platforms because they used to have 10.0.2, which
has wl_shell support. Now that stm32mp platforms use 13.0.1, use the 12.0.4
patch as reference and adapt it so it applies and builds correctly.

This makes it possible to run the LVGL demo on the ccmp25-dvk.

https://onedigi.atlassian.net/browse/DEL-9458

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>