tf-a-stm32mp: create symlinks to TF-A artifacts in subfolders
The following TF-A artifacts are deployed to subdirectories of
the image deploydir:
  - arm-trusted-firmware/tf-a-<platform>-<bootconfig>.stm32
  - arm-trusted-firmware/metadata-<platform>.bin
  - fip/fip-<platform>-<bootconfig>.bin
  - fip/fip-<platform>-ddr-<bootconfig>.bin

These binaries are not copied to the image deploy dir during
the regular do_deploy(), instead, they are deployed by script
tf_a_sysroot_populate() which is added to SYSROOT_PREPROCESS_FUNCS.

To follow this logic, change the previously wrong do_deploy:append
into a new function and append it also to SYSROOT_PREPROCESS_FUNCS,
so that it is called after the artifacts have really been deployed.

In the existing code, fix the paths and commands, which had some
errors.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DEL-9461