imx-digi-base.inc: add new imxviv override
This override was recently added to meta-freescale's graphical package recipes
to easily distinguish between i.MX platforms that use the Vivante driver and
those that use the Mali one. Without this change, a lot of these recipes are
broken, so sync with the latest changes in meta-freescale's imx-base.inc file.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>