x-linux-ai: recipes-samples: add recipe to manage the DCMIPP ISP
This recipe installs the dcmipp-isp-ctrl tool, an application that demonstrates
some of the features of the DCMIPP ISP blocks. It allows configuring
illuminance and auto-exposure to achieve a targeted level by adjusting both the
DCMIPP exposure block and the sensor gain.

This recipes is original from meta-st-openstlinux layer based on
openstlinux-6.6-yocto-scarthgap-mpu-v24.12.05 release.

https://onedigi.atlassian.net/browse/DEL-9486

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>