connectcore-demo-server: add delay after start demo server service
Since the ConnectCore demo server forks execution to start the demo server
Python script, it may not be fully initialized when the ConnectCore demo
example initializes the UI and attempts to retrieve information from the server
daemon. This commit introduces a delay to ensure the server is ready before
another service, such as the ConnectCore demo example, interacts with it.

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>