bootcount: skip resetting bootcount if bootlimit is not set
bootcount utility started by bootcount-init script was inconditionally setting
bootcount to 0 even if the functionality is disabled.

According to documentation, bootcount functionality is only enabled if bootlimit
variable is defined to a value greater than 0 in U-Boot environment.

In platforms like cc6ul, that use MCA NVRAM to store bootcount, this was corrupting
user NVRAM.

Version increased to 1.1.

Signed-off-by: Hector Bujanda <hector.bujanda@digi.com>